Adjusting Movement & Catchables
Modified the movement interface to delegate the application of the resulting velocity to the interface instead of handling it directly to work with both 2D and 3D rigidbodies.

Line spawner now focuses on 2D rigidbodies since the plan is for the fruit to be 2D. Same for all catchable related collision.

May need to remove the PlayerController movement interface and put it on its own FPS movement component.