Upgrading Catch Scripts
Catch game prototype was chosen, so its been upgraded from prototype.

Started working on separating some of the data to more easily reuse it and give access it. Associated catch action and colour is now used. For UI an extra wrapper is being worked on to potentially display on UI inputs being pressed and main action to clearly visualise perceived intent.

Line spawner now contains a list of actions that can be selected to set to the spawned catchable.