Updating UI & GameScene
Changing UI to use the new textures

BACKGROUND scene renamed to GameScene, it will now be considered the build scene.

The old FallingObjectsTest scene is now named TestGround and will be used for testing purposes.

Adjusted the trunk mesh materials by adding normal maps and slighly increasing the roughness to reduce reflection.

Imported mesh had to be changed to legacy to enable only affecting the trunk and not reintroducing emission issues.