Animations, Bouncer, Falling & Flipping
Player now has animations.

Added a animation parameter hasher to cache the parameter string ID.

Separated the bounce logic from the fail logic and replaced background colliders with bouncers.

Added a simple falling movement for the falling fruit.

Modified the movement interface to receive vector3 instead to be able to move in any direction.

Added UV flipping to the shader to be able to utilise the sprite renderer flip options.